re: #2836  support for asset deployment summary query
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@7879 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/asset/org.wcs.smart.asset.query/src/org/wcs/smart/asset/query/engine/AssetPagedObservationResult.java
@@ -474,7 +474,7 @@ public class AssetPagedObservationResult extends AssetPagedWaypointResult implem
 					updated = true;
 				}else if (newValue instanceof Double){
 					Double newDouble = (Double)newValue;
-					if (toUpdate.getNumberValue() != newDouble){
+					if (toUpdate.getNumberValue().doubleValue() != newDouble.doubleValue()){
 						toUpdate.setNumberValue(newDouble);
 						updated = true;
 					}
